Improve the connecting/disconnecting process
Allow easy access to manager instead of deciding which properties to expose.
Invoke the connect callback with an error when it fails to connect.
Ensure using a fresh GATT instance for each connection attempt.
Ensure you only call `.close` on a disconnected GATT instance.
Attempt to connect as BLE only if the API is supported.
Only retry connecting once.